Predicting with Linear Regression (x1+x2,y)
This file contains a supervised learning model developed in Python jupyter for the "Problem Set 1/b" of the CS229 course. The task involves predicting  using linear regression. The model utilizes two features, denoted as x1 and x2, to predict a single target variable, y. Additionally, HTML formatting is applied to enhance the visual presentation, such as adding colors to the headers.